Server/requirements.txt
flask
paho-mqtt
